requirements.txt
fastapi
uvicorn
crewai
crewai-tools
sqlalchemy
alembic
python-dotenv
markdown
hypothesis
pytest
httpx
